updateDebitAccount

+++ src/main/java/bankmicroservicesapp/entity/Account.java
@@ -11,6 +11,7 @@ import lombok.Setter;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Objects;
@@ -44,7 +45,7 @@ public class Account {
     private StatusAccount status;
 
     @Column(name = "balance")
-    private double balance;
+    private BigDecimal balance;
 
     @Column(name = "currency_code")
     @Enumerated(EnumType.STRING)
