+++ src/main/java/bankmicroservicesapp/entity/Transaction.java
@@ -1,16 +1,14 @@
 package bankmicroservicesapp.entity;
 
 import bankmicroservicesapp.entity.enums.TypeTransaction;
-import com.fasterxml.jackson.annotation.JsonIgnore;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 
-import java.time.LocalDate;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.Objects;
 import java.util.UUID;
@@ -35,7 +33,7 @@ public class Transaction {
     private TypeTransaction type;
 
     @Column(name = "amount")
-    private double amount;
+    private BigDecimal amount;
 
     @Column(name = "transaction_description")
     private String description;
